requirements.txt
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.2
streamlit==1.35.0
streamlit-authenticator==0.3.2
